feedback: rename cmd api package to client
