+++ cmd/cli/client/ucan/upload.go
@@ -13,7 +13,7 @@ import (
 	"github.com/storacha/go-ucanto/core/ipld/hash/sha256"
 	"github.com/storacha/go-ucanto/did"
 
-	"github.com/storacha/piri/cmd/api"
+	"github.com/storacha/piri/cmd/client"
 	"github.com/storacha/piri/pkg/config"
 )
 
@@ -45,7 +45,7 @@ func doUpload(cmd *cobra.Command, _ []string) error {
 		return fmt.Errorf("loading config: %w", err)
 	}
 
-	c, err := api.GetClient(cfg)
+	c, err := client.New(cfg)
 	if err != nil {
 		return fmt.Errorf("creating client: %w", err)
 	}
